requirements.txt
amqp==5.0.6
arabic-reshaper==2.1.3
asgiref==3.3.4
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.7
billiard==3.6.4.0
blessings==1.7
bpython==0.21
Brotli==1.0.9
cachetools==4.2.2
cairocffi==1.2.0
CairoSVG==2.5.2
celery==5.1.2
certifi==2021.5.30
cffi==1.14.5
channels==3.0.4
chardet==4.0.0
charset-normalizer==2.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorful==0.5.4
constantly==15.1.0
cryptography==3.4.7
cssselect2==0.4.1
curtsies==0.3.5
cwcwidth==0.1.4
cycler==0.10.0
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-countries==7.2.1
django-extensions==3.1.3
django-ipware==3.0.7
django-maintenance-mode==0.16.0
django-session-timeout==0.1.0
django-storages==1.11.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.4
et-xmlfile==1.1.0
Faker==8.10.3
fonttools==4.25.2
fpdf==1.7.2
future==0.18.2
google-api-core==1.31.1
google-api-python-client==2.15.0
google-auth==1.34.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.5
googleapis-common-protos==1.53.0
greenlet==1.1.0
gunicorn==20.1.0
html5lib==1.1
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
imbalanced-learn==0.8.0
imblearn==0.0
incremental==21.3.0
joblib==1.0.1
kiwisolver==1.3.1
kombu==5.1.0
lxml==4.6.3
matplotlib==3.4.2
numpy==1.21.1
oauthlib==3.1.1
openpyxl==3.0.7
packaging==21.0
pandas==1.3.1
pandas-datareader==0.10.0
pathtools==0.1.2
Pillow==8.3.1
prettyprinter==0.18.0
prompt-toolkit==3.0.19
protobuf==3.17.3
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pydyf==0.0.3
Pygments==2.9.0
PyHamcrest==2.0.2
PyJWT==2.1.0
PyMuPDF==1.18.15
pyOpenSSL==20.0.1
pyparsing==2.4.7
PyPDF2==1.26.0
pyphen==0.11.0
PyPika==0.48.8
python-bidi==0.4.2
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
pyxdg==0.27
reportlab==3.5.68
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
seaborn==0.11.1
service-identity==21.1.0
simplejson==3.17.3
six==1.16.0
sqlparse==0.4.1
stripe==2.60.0
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.1.0
timeago==1.0.15
tinycss2==1.1.0
toml==0.10.2
Twisted==20.3.0
txaio==21.2.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
wcwidth==0.2.5
weasyprint==53.0
webencodings==0.5.1
whitenoise==5.3.0
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zope.interface==5.4.0
zopfli==0.1.8
